requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Vector Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Embeddings & ML
openai==1.3.7
anthropic==0.7.8
sentence-transformers==2.2.2
cohere==4.37.0
numpy==1.24.3
scikit-learn==1.3.2

# Document Processing
pymupdf==1.23.8
marker-pdf==0.2.0
pytesseract==0.3.10
Pillow==10.1.0
python-multipart==0.0.6

# Text Processing
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2

# Caching & Message Queue
redis==5.0.1
celery==5.3.4
rabbitmq==3.13.0

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0

# HTTP & API
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Testing & Evaluation
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Jupyter for Analysis
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
